--- src/main/java/eu/comsode/unifiedviews/plugins/transformer/geonamesorgtontriples/GeonamesOrgToNTriplesConfig_V1.java
@@ -1,7 +1,0 @@
-package eu.comsode.unifiedviews.plugins.transformer.geonamesorgtontriples;
-
-/**
- */
-public class GeonamesOrgToNTriplesConfig_V1 {
-
-}
