+++ src/main/java/eu/comsode/unifiedviews/plugins/transformer/geonamesorgtordffile/Atts.java
@@ -14,7 +14,7 @@
  * implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package eu.comsode.unifiedviews.plugins.transformer.geonamesorgtontriples;
+package eu.comsode.unifiedviews.plugins.transformer.geonamesorgtordffile;
 
 import java.util.ArrayList;
 import java.util.Iterator;
