+++ src/main/java/eu/comsode/unifiedviews/plugins/transformer/geonamesorgtordffile/GeonamesOrgToRdfFileConfig_V1.java
@@ -1,0 +1,7 @@
+package eu.comsode.unifiedviews.plugins.transformer.geonamesorgtordffile;
+
+/**
+ */
+public class GeonamesOrgToRdfFileConfig_V1 {
+
+}
