+++ src/main/resources/resources.properties
@@ -1,1 +1,1 @@
-GeonamesOrgToNTriples.execute.exception = Error during conversion
+GeonamesOrgToRdfFile.execute.exception = Error during conversion
