+++ src/test/java/eu/comsode/unifiedviews/plugins/transformer/geonamesorgtordffile/GeonamesOrgToRdfFileTest.java
@@ -1,4 +1,4 @@
-package eu.comsode.unifiedviews.plugins.transformer.geonamesorgtontriples;
+package eu.comsode.unifiedviews.plugins.transformer.geonamesorgtordffile;
 
 import java.io.File;
 import java.io.InputStream;
@@ -11,18 +11,20 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import cz.cuni.mff.xrg.odcs.dpu.test.TestEnvironment;
+import eu.comsode.unifiedviews.plugins.transformer.geonamesorgtordffile.GeonamesOrgToRdfFile;
+import eu.comsode.unifiedviews.plugins.transformer.geonamesorgtordffile.GeonamesOrgToRdfFileConfig_V1;
 import eu.unifiedviews.dataunit.files.FilesDataUnit;
 import eu.unifiedviews.dataunit.files.WritableFilesDataUnit;
 import eu.unifiedviews.helpers.dataunit.files.FilesHelper;
 import eu.unifiedviews.helpers.dpu.test.config.ConfigurationBuilder;
 
-public class GeonamesOrgToNTriplesTest {
+public class GeonamesOrgToRdfFileTest {
     @Test
     public void testSmallFile() throws Exception {
-        GeonamesOrgToNTriplesConfig_V1 config = new GeonamesOrgToNTriplesConfig_V1();
+        GeonamesOrgToRdfFileConfig_V1 config = new GeonamesOrgToRdfFileConfig_V1();
 
         // Prepare DPU.
-        GeonamesOrgToNTriples dpu = new GeonamesOrgToNTriples();
+        GeonamesOrgToRdfFile dpu = new GeonamesOrgToRdfFile();
         dpu.configure((new ConfigurationBuilder()).setDpuConfiguration(config).toString());
 
         // Prepare test environment.
